Use tornado IOLoop instead of threading
In the real application the tornado IOLoop will be used to execute the
copying part of the staging service. However doing it this way has the
nice side effect of making the unit tests synchronous, and thus much
faster.

I'm leaving the test code with the assumption that it can be
asynchronous though, since in the future we might want some more
advanced testing that actually checks it in that way.